# platform/api-spec.yaml
openapi: 3.0.3
info:
  title: Sandbox Platform API
  description: |
    API for the Sandbox Platform — provisions isolated demo environments
    with AI-generated data. The backend manages projects, scenarios, and
    sandboxes. The frontend consumes this API.

    Base URL during development: http://localhost:4000/api
  version: 1.0.0

servers:
  - url: http://localhost:4000/api
    description: Local development backend

paths:
  /projects:
    get:
      summary: List all projects
      operationId: listProjects
      responses:
        "200":
          description: Array of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"

  /projects/{id}:
    get:
      summary: Get project with scenarios and sandboxes
      operationId: getProject
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project with related data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectWithDetails"
        "404":
          $ref: "#/components/responses/NotFound"

  /projects/{id}/scenarios:
    post:
      summary: Create a scenario under a project
      operationId: createScenario
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScenarioRequest"
      responses:
        "201":
          description: Created scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /sandboxes:
    post:
      summary: Create and provision a sandbox (async)
      description: |
        Creates a sandbox record and kicks off provisioning asynchronously.
        Returns immediately with status "provisioning". Poll GET /sandboxes/{id}
        for status updates.
      operationId: createSandbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSandboxRequest"
      responses:
        "201":
          description: Sandbox created, provisioning started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "400":
          $ref: "#/components/responses/BadRequest"

  /sandboxes/{id}:
    get:
      summary: Get sandbox status
      description: |
        Poll this endpoint every 2 seconds during provisioning to get
        status updates. Check `status` and `currentStep` fields.
      operationId: getSandbox
      parameters:
        - $ref: "#/components/parameters/SandboxId"
      responses:
        "200":
          description: Current sandbox state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Destroy a sandbox
      description: Stops Docker container, drops database, updates status.
      operationId: destroySandbox
      parameters:
        - $ref: "#/components/parameters/SandboxId"
      responses:
        "200":
          description: Sandbox destroyed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string

    SandboxId:
      name: id
      in: path
      required: true
      schema:
        type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"

  schemas:
    SchemaFormat:
      type: string
      enum: [prisma, sql]

    SandboxStatus:
      type: string
      enum: [provisioning, running, failed, destroyed]

    ProvisioningStep:
      type: string
      enum:
        - creating_database
        - applying_schema
        - generating_data
        - seeding_database
        - launching_app
        - waiting_for_ready
        - ready

    DemoUser:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string

    Project:
      type: object
      required: [id, name, dockerImage, schema, schemaFormat, appPort, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        dockerImage:
          type: string
        schema:
          type: string
        schemaFormat:
          $ref: "#/components/schemas/SchemaFormat"
        appPort:
          type: integer
          default: 3000
        createdAt:
          type: string
          format: date-time

    Scenario:
      type: object
      required: [id, projectId, name, prompt, demoUsers, featureFlags, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        prompt:
          type: string
        generatedSQL:
          type: string
          nullable: true
        demoUsers:
          type: array
          items:
            $ref: "#/components/schemas/DemoUser"
        featureFlags:
          type: object
          additionalProperties:
            type: boolean
        createdAt:
          type: string
          format: date-time

    Sandbox:
      type: object
      required: [id, projectId, scenarioId, subdomain, databaseName, status, statusMessage, expiresAt, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        scenarioId:
          type: string
        subdomain:
          type: string
        containerId:
          type: string
          nullable: true
        hostPort:
          type: integer
          nullable: true
        databaseName:
          type: string
        status:
          $ref: "#/components/schemas/SandboxStatus"
        statusMessage:
          type: string
        currentStep:
          $ref: "#/components/schemas/ProvisioningStep"
        url:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    ProjectWithDetails:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            scenarios:
              type: array
              items:
                $ref: "#/components/schemas/Scenario"
            sandboxes:
              type: array
              items:
                $ref: "#/components/schemas/Sandbox"

    CreateProjectRequest:
      type: object
      required: [name, dockerImage, schema, schemaFormat]
      properties:
        name:
          type: string
        dockerImage:
          type: string
        schema:
          type: string
        schemaFormat:
          $ref: "#/components/schemas/SchemaFormat"
        appPort:
          type: integer
          default: 3000

    CreateScenarioRequest:
      type: object
      required: [name, prompt]
      properties:
        name:
          type: string
        prompt:
          type: string
        demoUsers:
          type: array
          items:
            $ref: "#/components/schemas/DemoUser"
        featureFlags:
          type: object
          additionalProperties:
            type: boolean

    CreateSandboxRequest:
      type: object
      required: [projectId, scenarioId]
      properties:
        projectId:
          type: string
        scenarioId:
          type: string
        useWarmStart:
          type: boolean
          default: false

    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string